requirements.txt
flask
flask-cors
pyjwt
requests
gunicorn